requirements.txt
torch>=1.8.0
torchvision>=0.9.0
numpy>=1.24.2  # 최신 버전 유지, 1.23.0 제거
pyyaml>=6.0    # 최신 버전 유지, pyyaml>=5.3.1 제거
requests>=2.23.0
opencv-python-headless>=4.8.0.76  # headless 버전 유지, 일반 opencv-python 제거
Flask>=2.2.3
